edição de lider e grupo

+++ src/main/java/com/orbytum/api/repository/UsuarioRepository.java
@@ -1,13 +1,14 @@
 package com.orbytum.api.repository;
 
-import com.orbytum.api.models.entity.Usuario;
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
+import com.orbytum.api.models.entity.Usuario;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 
     Optional<Usuario> findByEmail(String email);
 }
